+++ experimental/packages/opentelemetry-instrumentation/src/platform/node/RequireInTheMiddleSingleton.ts
@@ -15,6 +15,7 @@
  */
 
 import * as RequireInTheMiddle from 'require-in-the-middle';
+import * as ImportInTheMiddle from 'import-in-the-middle';
 import * as path from 'path';
 import { ModuleNameTrie, ModuleNameSeparator } from './ModuleNameTrie';
 
@@ -22,6 +23,12 @@ export type Hooked = {
   moduleName: string
   onRequire: RequireInTheMiddle.OnRequireFn
 };
+
+/**
+ * We are forced to re-type there because ImportInTheMiddle is exported as normal CJS
+ * in the JS files but transpiled ESM (with a default export) in its typing.
+ */
+ const ESMHook = ImportInTheMiddle as unknown as typeof ImportInTheMiddle.default;
 
 /**
  * Whether Mocha is running in this process
@@ -35,12 +42,12 @@ const isMocha = ['afterEach','after','beforeEach','before','describe','it'].ever
 });
 
 /**
- * Singleton class for `require-in-the-middle`
+ * Singleton class for `require-in-the-middle` and `import-in-the-middle`
  * Allows instrumentation plugins to patch modules with only a single `require` patch
- * WARNING: Because this class will create its own `require-in-the-middle` (RITM) instance,
+ * WARNING: Because this class will create its own RITM and IITM instance,
  * we should minimize the number of new instances of this class.
  * Multiple instances of `@opentelemetry/instrumentation` (e.g. multiple versions) in a single process
- * will result in multiple instances of RITM, which will have an impact
+ * will result in multiple instances of RITM/ITTM, which will have an impact
  * on the performance of instrumentation hooks being applied.
  */
 export class RequireInTheMiddleSingleton {
@@ -52,23 +59,23 @@ export class RequireInTheMiddleSingleton {
   }
 
   private _initialize() {
-    RequireInTheMiddle(
-      // Intercept all `require` calls; we will filter the matching ones below
-      null,
-      { internals: true },
-      (exports, name, basedir) => {
-        // For internal files on Windows, `name` will use backslash as the path separator
-        const normalizedModuleName = normalizePathSeparators(name);
+    // 
+    const onHook = (exports: any, name: string, basedir: string | undefined | void) => {
+      // For internal files on Windows, `name` will use backslash as the path separator
+      const normalizedModuleName = normalizePathSeparators(name);
 
-        const matches = this._moduleNameTrie.search(normalizedModuleName, { maintainInsertionOrder: true });
+      const matches = this._moduleNameTrie.search(normalizedModuleName, { maintainInsertionOrder: true });
 
-        for (const { onRequire } of matches) {
-          exports = onRequire(exports, name, basedir);
-        }
-
-        return exports;
+      for (const { onRequire } of matches) {
+        exports = onRequire(exports, name, basedir ? basedir : undefined);
       }
-    );
+
+      return exports;
+    }
+    // Intercept all `require` calls; we will filter the matching ones below
+    RequireInTheMiddle(null, { internals: true }, onHook);
+    // We can give no module to patch but this signature isn't exposed in typings
+    new ESMHook(null as any, { internals: true }, onHook)
   }
 
   /**